Renamed configErrors to errors
