+++ src/core/config/errors.ts
